+++ packages/daisyui/src/templates/FieldTemplate/FieldTemplate.tsx
@@ -38,7 +38,6 @@ export default function FieldTemplate<
     required,
     rawErrors,
     rawHelp,
-    description,
     rawDescription,
     hidden,
     onChange,
@@ -59,20 +58,7 @@ export default function FieldTemplate<
           </span>
         </label>
       )}
-      {isCheckbox ? (
-        <div className='form-control'>
-          {description}
-          <label className='label cursor-pointer justify-start'>
-            <div className='mr-2'>{children}</div>
-            <span className='label-text'>
-              {label}
-              {required && <span className='text-error ml-1'>*</span>}
-            </span>
-          </label>
-        </div>
-      ) : (
-        children
-      )}
+      {children}
       {errors}
       {help}
     </div>